chore: update to libdns/libdns v0.2.2 - fix build errors with caddy v2.8.0+ (#3)
* chore: update to libdns/libdns v0.2.2

* cleanup: fix build error with libdns v0.2.2

+++ models.go
@@ -55,7 +55,7 @@ func (r daRecord) libdnsRecord(zone string) (libdns.Record, error) {
 			return record, fmt.Errorf("failed to parse MX priority for %v: %v", r.Name, err)
 		}
 
-		record.Priority = priority
+		record.Priority = uint(priority)
 		record.Value = fmt.Sprintf("%v.%v", splits[1], zone)
 	case "SRV":
 		return record, ErrUnsupported
